requirements.txt
boto==2.36.0
requests==2.6.0
argparse==1.3.0
python-firebase==1.2
wsgiref==0.1.2
github3.py==0.9.3
